requirements/dev.txt
# Install requirements for testing
# This is used for development and for Travis CI

# Production requirements
-r ./prod.txt

# Testing
nose
coverage
coveralls
flask-testing
factory-boy